// packages/adapter-telegram/src/types.ts
export interface PassportData {
    data: EncryptedPassportElement[];
    credentials: EncryptedCredentials;
}

export interface Update {
    updateId: number;
    message?: Message;
    editedMessage?: Message;
    channelPost?: Message;
    editedChannelPost?: Message;
    inlineQuery?: InlineQuery;
    chosenInlineResult?: ChosenInlineResult;
    callbackQuery?: CallbackQuery;
    shippingQuery?: ShippingQuery;
    preCheckoutQuery?: PreCheckoutQuery;
}

export interface WebhookInfo {
    url: string;
    hasCustomCertificate: boolean;
    pendingUpdateCount: number;
    lastErrorDate?: number;
    lastErrorMessage?: string;
    maxConnections?: number;
    allowedUpdates?: string[];
}

export interface User {
    id: number;
    isBot: boolean;
    firstName: string;
    lastName?: string;
    username?: string;
    languageCode?: string;
}

export interface Chat {
    id: number;
    type: ChatType;
    title?: string;
    username?: string;
    firstName?: string;
    lastName?: string;
    photo?: ChatPhoto;
    description?: string;
    inviteLink?: string;
    pinnedMessage?: Message;
    permissions?: ChatPermissions;
    canSetStickerSet?: boolean;
    stickerSetName?: string;
    /**
     * @deprecated since version Telegram Bot API 4.4 - July 29, 2019
     */
    allMembersAreAdministrators?: boolean;
}

export interface Message {
    messageId: number;
    from?: User;
    date: number;
    chat: Chat;
    forwardFrom?: User;
    forwardFromChat?: Chat;
    forwardFromMessageId?: number;
    forwardSignature?: string;
    forwardSenderName?: string;
    forwardDate?: number;
    replyToMessage?: Message;
    editDate?: number;
    mediaGroupId?: string;
    authorSignature?: string;
    text?: string;
    entities?: MessageEntity[];
    captionEntities?: MessageEntity[];
    audio?: Audio;
    document?: Document;
    animation?: Animation;
    game?: Game;
    photo?: PhotoSize[];
    sticker?: Sticker;
    video?: Video;
    voice?: Voice;
    videoNote?: VideoNote;
    caption?: string;
    contact?: Contact;
    location?: Location;
    venue?: Venue;
    poll?: Poll;
    newChatMembers?: User[];
    leftChatMember?: User;
    newChatTitle?: string;
    newChatPhoto?: PhotoSize[];
    deleteChatPhoto?: boolean;
    groupChatCreated?: boolean;
    supergroupChatCreated?: boolean;
    channelChatCreated?: boolean;
    migrateToChatId?: number;
    migrateFromChatId?: number;
    pinnedMessage?: Message;
    invoice?: Invoice;
    successfulPayment?: SuccessfulPayment;
    connectedWebsite?: string;
    passportData?: PassportData;
    replyMarkup?: InlineKeyboardMarkup;
}

export interface MessageEntity {
    type: MessageEntityType;
    offset: number;
    length: number;
    url?: string;
    user?: User;
}

export interface FileBase {
    fileId: string;
    fileSize?: number;
}

export interface PhotoSize extends FileBase {
    width: number;
    height: number;
}

export interface Audio extends FileBase {
    duration: number;
    performer?: string;
    title?: string;
    mimeType?: string;
    thumb?: PhotoSize;
}

export interface Document extends FileBase {
    thumb?: PhotoSize;
    fileName?: string;
    mimeType?: string;
}

export interface Video extends FileBase {
    width: number;
    height: number;
    duration: number;
    thumb?: PhotoSize;
    mimeType?: string;
}

export interface Voice extends FileBase {
    duration: number;
    mimeType?: string;
}
// ...
export interface VideoNote extends FileBase {
    length: number;
    duration: number;
    thumb?: PhotoSize;
}

export interface Contact {
    phoneNumber: string;
    firstName: string;
    lastName?: string;
    userId?: number;
    vcard?: string;
}

export interface Location {
    longitude: number;
    latitude: number;
}

export interface Venue {
    location: Location;
    title: string;
    address: string;
    foursquareId?: string;
    foursquareType?: string;
}

export type PollType = 'regular' | 'quiz';

export interface PollAnswer {
    pollId: string;
    user: User;
    optionIds: number[];
}

export interface PollOption {
    text: string;
    voterCount: number;
}

export interface Poll {
    id: string;
    question: string;
    options: PollOption[];
    isClosed: boolean;
    isAnonymous: boolean;
    allowsMultipleAnswers: boolean;
    type: PollType;
    totalVoterCount: number;
}
// ...
export interface InlineKeyboardMarkup {
    inlineKeyboard: InlineKeyboardButton[][];
}

export interface InlineKeyboardButton {
    text: string;
    url?: string;
    loginUrl?: LoginUrl;
    callbackData?: string;
    switchInlineQuery?: string;
    switchInlineQueryCurrentChat?: string;
    callbackGame?: CallbackGame;
    pay?: boolean;
}

export interface LoginUrl {
    url: string;
    forwardText?: string;
    botUsername?: string;
    requestWriteAcces?: boolean;
}

export interface CallbackQuery {
    id: string;
    from: User;
    message?: Message;
    inlineMessageId?: string;
    chatInstance: string;
    data?: string;
    gameShortName?: string;
}
// ...
export interface Sticker {
    fileId: string;
    fileUniqueId: string;
    isAnimated: boolean;
    width: number;
    height: number;
    thumb?: PhotoSize;
    emoji?: string;
    setName?: string;
    maskPosition?: MaskPosition;
    fileSize?: number;
}

export interface StickerSet {
    name: string;
    title: string;
    containsMasks: boolean;
    stickers: Sticker[];
}

export interface MaskPosition {
    point: string;
    xShift: number;
    yShift: number;
    scale: number;
}

export interface InlineQuery {
    id: string;
    from: User;
    location?: Location;
    query: string;
    offset: string;
}
// ...
export interface ChosenInlineResult {
    resultId: string;
    from: User;
    location?: Location;
    inlineMessageId?: string;
    query: string;
}

export interface ResponseParameters {
    migrateToChatId?: number;
    retryAfter?: number;
}

export interface LabeledPrice {
    label: string;
    amount: number;
}

export interface Invoice {
    title: string;
    description: string;
    startParameter: string;
    currency: string;
    totalAmount: number;
}

export interface ShippingAddress {
    countryCode: string;
    state: string;
    city: string;
    streetLine1: string;
    streetLine2: string;
    postCode: string;
}

export interface OrderInfo {
    name?: string;
    phoneNumber?: string;
    email?: string;
    shippingAddress?: ShippingAddress;
}

export interface ShippingOption {
    id: string;
    title: string;
    prices: LabeledPrice[];
}

export interface SuccessfulPayment {
    currency: string;
    totalAmount: number;
    invoicePayload: string;
    shippingOptionId?: string;
    orderInfo?: OrderInfo;
    telegramPaymentChargeId: string;
    providerPaymentChargeId: string;
}

export interface ShippingQuery {
    id: string;
    from: User;
    invoicePayload: string;
    shippingAddress: ShippingAddress;
}

export interface PreCheckoutQuery {
    id: string;
    from: User;
    currency: string;
    totalAmount: number;
    invoicePayload: string;
    shippingOptionId?: string;
    orderInfo?: OrderInfo;
}

export interface Game {
    title: string;
    description: string;
    photo: PhotoSize[];
    text?: string;
    textEntities?: MessageEntity[];
    animation?: Animation;
}

export interface Animation extends FileBase {
    width: number;
    height: number;
    duration: number;
    thumb?: PhotoSize;
    fileName?: string;
    mimeType?: string;
}